grok-voice-agent/requirements.txt
# LiveKit Agents with xAI plugin
livekit-agents[xai]>=1.0.0

# Environment variables
python-dotenv>=1.0.0
